move BlockwiseCreateArray into layers.py
